Extract hook useTasks from TaskManager
