+++ src/components/TaskManager.tsx
@@ -1,55 +1,22 @@
 import React, { useEffect, useState } from "react";
 
-import { Item } from "../constants/Item";
-
 import TaskItem from "./TaskItem";
 import FilterButton from "./shared/FilterButton";
+import useTasks from "../hooks/useTasks";
 
 const TaskManager = () => {
-  const [tasks, setTasks] = useState<Item[]>([
-    { id: 1, title: "Buy groceries", completed: false },
-    { id: 2, title: "Clean the house", completed: true },
-  ]);
-  const [filter, setFilter] = useState("all");
-  const [newTask, setNewTask] = useState<string>("");
-  const [lastUpdate, setLastUpdate] = useState<Date>(new Date());
-
-  useEffect(() => setLastUpdate(new Date()), [tasks]);
-
-  const filteredTasks = tasks.filter((task) => {
-    if (filter === "completed") return task.completed;
-    if (filter === "pending") return !task.completed;
-    return true;
-  });
-
-  const handleAddTask = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (newTask.trim() === "") return;
-
-    const maxId = tasks.reduce((max, task) => Math.max(max, task.id), 0)
-
-    const newTaskObj:Item = {
-      id: maxId + 1,
-      title: newTask,
-      completed: false,
-    };
-
-    setTasks([...tasks, newTaskObj]);
-    setNewTask("");
-  };
-
-  const handleDeleteTask = (id: number) => {
-    setTasks(tasks.filter((task) => task.id !== id));
-  };
-
-  const toggleTaskCompletion = (id: number) => {
-    const task = tasks.find((task) => task.id === id);
-
-    if (!task) return;
-
-    setLastUpdate(new Date());
-    task.completed = !task.completed;
-  };
+  const {
+    filter,
+    filters,
+    handleAddTask,
+    handleDeleteTask,
+    lastUpdate,
+    newTask,
+    setFilter,
+    setNewTask,
+    tasks,
+    toggleTaskCompletion,
+  } = useTasks();
 
   return (
     <div className="container mx-auto bg-white p-4 rounded shadow">
@@ -66,12 +33,10 @@ const TaskManager = () => {
         </button>
       </form>
       <div className="flex justify-around mb-4">
-        <FilterButton filterKey={"all"} currentFilter={filter} setFilter={setFilter} />
-        <FilterButton filterKey={"completed"} currentFilter={filter} setFilter={setFilter} />
-        <FilterButton filterKey={"pending"} currentFilter={filter} setFilter={setFilter} />
+        {filters.map((filterKey) => <FilterButton key={filterKey} filterKey={filterKey} currentFilter={filter} setFilter={setFilter} />)}
       </div>
       <ul>
-        {filteredTasks.map((task) => (
+        {tasks.map((task) => (
           <TaskItem
             key={task.id}
             task={task}
